调整考试也被update.lua控制

+++ main.lua
@@ -10,19 +10,21 @@ if package.config:sub(1, 1) == "\\" then -- Windows
     binary_path = "client.exe"
 end
 
--- 如果主目录有 $binary_path.new，就用这个文件替换掉 $binary_path
-local new_path = binary_path .. ".new"
-local f = io.open(new_path, "r")
-if f then
-    f:close()
-    -- 存在 .new 文件，进行替换
-    os.remove(binary_path)
-    os.rename(new_path, binary_path)
-end
+if apply_upgrade then
+    -- 如果主目录有 $binary_path.new，就用这个文件替换掉 $binary_path
+    local new_path = binary_path .. ".new"
+    local f = io.open(new_path, "r")
+    if f then
+        f:close()
+        -- 存在 .new 文件，进行替换
+        os.remove(binary_path)
+        os.rename(new_path, binary_path)
+    end
 
--- 运行 $binary_path --quiz，强等待。
-local cmd = string.format('"%s" --quiz', binary_path)
-local ret = os.execute(cmd)
+    -- 运行 $binary_path --quiz，强等待。
+    local cmd = string.format('"%s" --quiz', binary_path)
+    local ret = os.execute(cmd)
+end
 
 local function check_update_async()
     local hash_file = io.open("current_version_commit_hash.txt", "r")
